Added ACL

+++ resources/views/blog/post.blade.php
@@ -9,14 +9,18 @@
     <button class="btn btn-primary" onclick="history.go(-1)">
         « Back
     </button>
-    <a href="/blog/{{ $post->id }}/edit">
-        <button class="btn btn-info">
-            Edit
-        </button>
-    </a>
-    <a href="/blog/{{ $post->id }}/delete">
-        <button class="btn btn-danger" href="/blog/{{ $post->id }}/delete">
-            Delete
-        </button>
-    </a>
+    @if(Auth::check())
+        @if(Auth::user()->role_id==\App\Role::ADMIN)
+            <a href="/blog/{{ $post->id }}/edit">
+                <button class="btn btn-info">
+                    Edit
+                </button>
+            </a>
+            <a href="/blog/{{ $post->id }}/delete">
+                <button class="btn btn-danger" href="/blog/{{ $post->id }}/delete">
+                    Delete
+                </button>
+            </a>
+        @endif
+    @endif
 @endsection
